# packages/contracts/openapi-v2.yaml
openapi: 3.1.0
info:
  title: Wuji vNext API
  version: 2.0.0
  description: >-
    Authoritative external wire contract for the isolated Wuji vNext runtime.
    Domain services are supplied by later stages; this document fixes their
    authenticated HTTP and worker boundaries.
servers:
  - url: /
    description: Same-origin Wuji vNext API
security:
  - BearerAuth: []
...
paths:
  /api/v2/tasks:
    get:
      operationId: listTasksV2
      summary: List currently authorized tasks in a project
      tags: [Tasks]
      parameters:
        - {name: project_id, in: query, required: true, schema: {type: string, minLength: 1, maxLength: 256}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
        - {name: cursor, in: query, schema: {type: string, minLength: 1, maxLength: 4096}}
      responses:
        '200':
          description: Authorized task page with no hidden count
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TaskList'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '404': {$ref: '#/components/responses/NotFoundOrForbidden'}
        '422': {$ref: '#/components/responses/InvalidSchema'}
    post:
      operationId: createTaskV2
      summary: Create a non-running vNext task
      tags: [Tasks]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created in the non-running ready state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskView'
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/InvalidSchema'}
  /api/v2/projects/{project_id}/task-options:
    get:
      operationId: getTaskOptionsV2
      summary: Read published options for an authorized creator
      tags: [Tasks]
      parameters:
        - {name: project_id, in: path, required: true, schema: {type: string, minLength: 1, maxLength: 256}}
      responses:
        '200':
          description: Redacted published options without credentials
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TaskOptions'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '404': {$ref: '#/components/responses/NotFoundOrForbidden'}
  /api/v2/tasks/{task_id}:
    get:
      operationId: getTaskV2
      summary: Read the current authorized task state
      tags: [Tasks]
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Current task state; accepted commands do not prove process exit
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TaskView'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '404': {$ref: '#/components/responses/NotFoundOrForbidden'}
  /api/v2/tasks/{task_id}/overview:
    get:
      operationId: getTaskOverviewV2
      summary: Read the bounded current task overview from canonical records
      tags: [Tasks]
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200': {description: Canonical task overview without model-generated summaries, content: {application/json: {schema: {$ref: '#/components/schemas/TaskOverviewV1'}}}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '404': {$ref: '#/components/responses/NotFoundOrForbidden'}
  /api/v2/tasks/{task_id}/command-inventory:
    get:
      operationId: getTaskCommandInventoryV1
      parameters:
        - {name: task_id, in: path, required: true, schema: {type: string}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 50}}
        - {name: after, in: query, schema: {type: string, minLength: 1, maxLength: 256}}
      responses:
        '200': {description: Bounded Task command inventory, content: {application/json: {schema: {$ref: '#/components/schemas/TaskCommandInventoryV1'}}}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '404': {$ref: '#/components/responses/NotFoundOrForbidden'}
  /api/v2/tasks/{task_id}/publications:
    get:
      operationId: getTaskPublicationsV1
      parameters:
        - {name: task_id, in: path, required: true, schema: {type: string}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 50}}
        - {name: after, in: query, schema: {type: string, minLength: 1, maxLength: 256}}
      responses:
        '200': {description: Bounded immutable workspace publications, content: {application/json: {schema: {$ref: '#/components/schemas/TaskPublicationPageV1'}}}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '404': {$ref: '#/components/responses/NotFoundOrForbidden'}
  /api/v2/tasks/{task_id}/activity:
    get:
      operationId: getTaskActivityV2
      summary: Read bounded canonical task activity; cursor and after_cursor are mutually exclusive
      tags: [Tasks]
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
        - {name: cursor, in: query, required: false, schema: {type: string, minLength: 1, maxLength: 4096}, description: Opaque task/filter-bound cursor for older history}
        - {name: after_cursor, in: query, required: false, schema: {type: string, minLength: 1, maxLength: 4096}, description: Opaque task/filter-bound cursor for newer incremental events}
        - {name: importance, in: query, required: false, schema: {type: string, enum: [key, all], default: key}}
        - {name: category, in: query, required: false, schema: {$ref: '#/components/schemas/TaskActivityCategory'}}
        - {name: work_item_id, in: query, required: false, schema: {type: string, minLength: 1, maxLength: 256}}
      responses:
        '200': {description: Authorized canonical activity page, content: {application/json: {schema: {$ref: '#/components/schemas/TaskActivityPageV1'}}}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '404': {$ref: '#/components/responses/NotFoundOrForbidden'}
        '422': {$ref: '#/components/responses/InvalidSchema'}
  /api/v2/tasks/{task_id}/readiness:
    get:
      operationId: getTaskReadinessV2
      summary: Read static readiness without contacting a target or model
      tags: [Tasks]
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Configuration facts and explicit unknown checks
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ReadinessReport'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '404': {$ref: '#/components/responses/NotFoundOrForbidden'}
...
  /api/v2/tasks/{task_id}/capture-sessions:
    get:
      operationId: listRuntimeCaptureSessionsV2
      summary: List authorized Runtime capture sessions for a Task
      tags: [Knowledge]
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200': {description: Bounded capture session inventory, content: {application/json: {schema: {$ref: '#/components/schemas/RuntimeCaptureSessionPageV1'}}}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '404': {$ref: '#/components/responses/NotFoundOrForbidden'}
  /api/v2/tasks/{task_id}/capture-sessions/{capture_session_id}/items:
    get:
      operationId: listRuntimeCaptureItemsV2
      summary: Page through authorized Runtime capture items
      tags: [Knowledge]
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - {name: capture_session_id, in: path, required: true, schema: {type: string, minLength: 1, maxLength: 256}}
        - {name: after, in: query, required: false, schema: {type: integer, minimum: 0, maximum: 1000000, default: 0}}
        - {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 100, default: 50}}
      responses:
        '200': {description: Bounded capture item page, content: {application/json: {schema: {$ref: '#/components/schemas/RuntimeCaptureItemPageV1'}}}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '404': {$ref: '#/components/responses/NotFoundOrForbidden'}
        '422': {$ref: '#/components/responses/InvalidSchema'}
  /api/v2/tasks/{task_id}/capture-sessions/{capture_session_id}/items/{item_seq}/parts/{part}:
    get:
      operationId: downloadRuntimeCapturePartV2
      summary: Download one authorized sealed Runtime capture part
      tags: [Knowledge]
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - {name: capture_session_id, in: path, required: true, schema: {type: string, minLength: 1, maxLength: 256}}
        - {name: item_seq, in: path, required: true, schema: {type: integer, minimum: 1, maximum: 1000000}}
        - {name: part, in: path, required: true, schema: {type: string, pattern: '^[a-z][a-z0-9_.-]{0,127}$'}}
      responses:
        '200':
          description: Exact sealed capture part bytes
          headers:
            Digest: {schema: {type: string}}
          content:
            application/octet-stream:
              schema: {type: string, format: binary}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '404': {$ref: '#/components/responses/NotFoundOrForbidden'}
        '422': {$ref: '#/components/responses/InvalidSchema'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 256}
...
    TaskId:
      name: task_id
      in: path
      required: true
      schema: {type: string, minLength: 1}
...
  responses:
...
    Unauthenticated:
      description: Missing or invalid authenticated identity
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
...
    NotFoundOrForbidden:
      description: Object is absent or hidden by the current read scope
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
    Conflict:
      description: Version, execution, digest, or operation state conflicts
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
...
    InvalidSchema:
      description: Request does not satisfy the versioned input contract
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
...
    Unavailable:
      description: Required runtime capability is unavailable
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
  schemas:
...
    RevisionString:
      type: string
      pattern: '^(0|[1-9][0-9]*)$'
...
    ApiSchemaVersion:
      type: string
      enum: [wuji.api.v2]
...
    TaskDesired:
      type: string
      enum: [run, pause, cancel, finish]
    TaskObserved:
      type: string
      enum: [ready, running, quiescing, paused, reconciling, closed]
    CloseTrigger:
      type: string
      enum: [goal_satisfied, user_cancel, budget_exhausted, time_limit, no_progress, system_failure, operator_finish]
    ResultOutcome:
      type: string
      enum: [complete, partial, inconclusive, not_assessed]
...
    TaskCommandName:
      type: string
      enum: [start, pause, resume, cancel, finish]
...
    ScenarioKind:
      type: string
      enum: [ctf, web_single, comprehensive, adversary_emulation, code_audit]
...
    RuntimeCaptureItemPageV1:
      type: object
      additionalProperties: false
      required: [schema_version, capture_session_id, items, next_item_seq]
      properties:
        schema_version: {type: string, const: wuji.runtime-capture-items.v1}
        capture_session_id: {type: string, minLength: 1, maxLength: 256}
        items:
          type: array
          maxItems: 100
          items: {$ref: '#/components/schemas/RuntimeCaptureItemViewV1'}
        next_item_seq:
          oneOf:
            - {type: integer, minimum: 1, maximum: 1000000}
            - {type: 'null'}
    RuntimeCaptureSessionPageV1:
      type: object
      additionalProperties: false
      required: [schema_version, sessions]
      properties:
        schema_version: {type: string, const: wuji.runtime-capture-sessions.v1}
        sessions:
          type: array
          maxItems: 100
          items: {$ref: '#/components/schemas/RuntimeCaptureSessionV1'}
...
    GoalContractInput:
      type: object
      additionalProperties: false
      required: [text, criteria]
      properties:
        text: {type: string, minLength: 1, maxLength: 65536}
        criteria:
          type: array
          minItems: 1
          maxItems: 256
          items: {$ref: '#/components/schemas/GoalCriterionInput'}
    AuthorizationScopeEntry:
      type: object
      additionalProperties: false
      required: [host, protocol, port]
      properties:
        host: {type: string, minLength: 1, maxLength: 253}
        protocol: {type: string, enum: [http, https]}
        port: {type: integer, minimum: 1, maximum: 65535}
    MoneyBudget:
      type: object
      additionalProperties: false
      required: [amount, currency]
      properties:
        amount: {$ref: '#/components/schemas/DecimalString'}
        currency: {type: string, enum: [USD]}
...
    TaskGoalOverviewV1:
      type: object
      additionalProperties: false
      required: [text, criteria]
      properties:
        text: {type: string, minLength: 1, maxLength: 65536}
        criteria:
          type: array
          minItems: 1
          maxItems: 256
          items: {$ref: '#/components/schemas/TaskCriterionOverviewV1'}
...
    TaskFindingSummaryV1:
      type: object
      additionalProperties: false
      required: [claim_ref, text, kind, grounding_state, evidence_state, applicability_state, supporting_refs, opposing_refs, updated_at]
      properties:
        claim_ref: {$ref: '#/components/schemas/KnowledgeRef'}
        text: {type: string, minLength: 1, maxLength: 32768}
        kind: {$ref: '#/components/schemas/ClaimKind'}
        grounding_state: {$ref: '#/components/schemas/GroundingState'}
        evidence_state: {$ref: '#/components/schemas/EvidenceState'}
        applicability_state: {$ref: '#/components/schemas/ApplicabilityState'}
        supporting_refs:
          type: array
          maxItems: 32
          items: {$ref: '#/components/schemas/KnowledgeRef'}
        opposing_refs:
          type: array
          maxItems: 32
          items: {$ref: '#/components/schemas/KnowledgeRef'}
        updated_at: {type: string, format: date-time}
...
    TaskOverviewV1:
      type: object
      additionalProperties: false
      required: [schema_version, task_id, observed_at, status_message, goal, budget, current_work, latest_findings, artifacts, workspace_capabilities, runtime, technical]
      properties:
        schema_version: {type: string, const: wuji.task-overview.v1}
        task_id: {type: string, minLength: 1, maxLength: 256}
        observed_at: {type: string, format: date-time}
        status_message: {type: string, maxLength: 512}
        goal: {$ref: '#/components/schemas/TaskGoalOverviewV1'}
        budget: {$ref: '#/components/schemas/MoneyBudget'}
        current_work:
          type: array
          maxItems: 5
          items: {$ref: '#/components/schemas/TaskWorkSummaryV1'}
        latest_findings:
          type: array
          maxItems: 5
          items: {$ref: '#/components/schemas/TaskFindingSummaryV1'}
        artifacts:
          type: array
          maxItems: 100
          items: {$ref: '#/components/schemas/TaskArtifactSummaryV1'}
        workspace_capabilities: {$ref: '#/components/schemas/TaskWorkspaceCapabilitiesV1'}
        runtime: {$ref: '#/components/schemas/TaskRuntimeOverviewV1'}
        technical: {$ref: '#/components/schemas/TaskTechnicalDetailsV1'}
...
    TaskCommandInventoryV1:
      type: object
      additionalProperties: false
      required: [schema_version, items, next_after]
      properties:
        schema_version: {type: string, const: wuji.command-inventory.v1}
        items:
          type: array
          maxItems: 100
          items: {$ref: '#/components/schemas/TaskCommandItemV1'}
        next_after:
          oneOf:
            - {type: string, minLength: 1, maxLength: 256}
            - {type: 'null'}
...
    TaskPublicationPageV1:
      type: object
      additionalProperties: false
      required: [schema_version, items, next_after]
      properties:
        schema_version: {type: string, const: wuji.task-publications.v1}
        items:
          type: array
          maxItems: 100
          items: {$ref: '#/components/schemas/TaskPublicationItemV1'}
        next_after:
          oneOf:
            - {type: string, minLength: 1, maxLength: 256}
            - {type: 'null'}
...
    TaskActivityCategory:
      type: string
      enum: [task, work, finding, evidence, input, completion, tool]
...
    TaskActivityPageV1:
      type: object
      additionalProperties: false
      required: [schema_version, task_id, observed_at, latest_cursor, items, next_cursor]
      properties:
        schema_version: {type: string, const: wuji.task-activity.v1}
        task_id: {type: string, minLength: 1, maxLength: 256}
        observed_at: {type: string, format: date-time}
        latest_cursor: {type: string, minLength: 1, maxLength: 4096}
        items:
          type: array
          maxItems: 100
          items: {$ref: '#/components/schemas/TaskActivityItemV1'}
        next_cursor:
          oneOf:
            - {type: string, minLength: 1, maxLength: 4096}
            - {type: 'null'}
    TaskCreate:
      type: object
      additionalProperties: false
      required: [schema_version, project_id, name, scenario, goal, authorization_scope, authorization_expires_at, model_profile_ref, runtime_profile_ref, budget]
      properties:
        schema_version: {$ref: '#/components/schemas/ApiSchemaVersion'}
        project_id: {type: string, minLength: 1, maxLength: 256}
        name: {type: string, minLength: 1, maxLength: 256}
        scenario: {$ref: '#/components/schemas/ScenarioKind'}
        goal: {$ref: '#/components/schemas/GoalContractInput'}
        authorization_scope:
          type: array
          minItems: 1
          maxItems: 1024
          items: {$ref: '#/components/schemas/AuthorizationScopeEntry'}
        authorization_expires_at: {type: string, format: date-time}
        entry_points:
          type: array
          minItems: 1
          maxItems: 16
          uniqueItems: true
          description: Explicit URLs preserving path and non-sensitive query; each origin must be authorized by the confirmed scope.
          items: {type: string, minLength: 1, maxLength: 2048}
        external_analysis_approved:
          type: boolean
          description: Creator confirmation that the authorized task material may be sent to the selected external model. Missing approval never permits external analysis.
        model_profile_ref: {type: string, minLength: 1, maxLength: 256}
        runtime_profile_ref: {type: string, minLength: 1, maxLength: 256}
        explore_concurrency:
          type: integer
          minimum: 1
          maximum: 256
          description: Task-selected Explore concurrency. When omitted, the published RuntimeProfile default applies.
        budget: {$ref: '#/components/schemas/MoneyBudget'}
    TaskView:
      type: object
      additionalProperties: false
      required: [task_id, tenant_id, project_id, version, name, scenario, desired_state, observed_state, goal_revision, execution_epoch, allowed_actions]
      properties:
        task_id: {type: string, minLength: 1}
        tenant_id: {type: string, minLength: 1}
        project_id: {type: string, minLength: 1}
        version: {$ref: '#/components/schemas/RevisionString'}
        name: {type: string, minLength: 1}
        scenario: {$ref: '#/components/schemas/ScenarioKind'}
        desired_state: {$ref: '#/components/schemas/TaskDesired'}
        observed_state: {$ref: '#/components/schemas/TaskObserved'}
        goal_revision: {$ref: '#/components/schemas/RevisionString'}
        execution_epoch: {$ref: '#/components/schemas/RevisionString'}
        activated_at:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
        close_trigger:
          oneOf:
            - $ref: '#/components/schemas/CloseTrigger'
            - type: 'null'
        result_outcome:
          oneOf:
            - $ref: '#/components/schemas/ResultOutcome'
            - type: 'null'
        allowed_actions:
          type: array
          uniqueItems: true
          items: {$ref: '#/components/schemas/TaskCommandName'}
    TaskList:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items: {type: array, maxItems: 100, items: {$ref: '#/components/schemas/TaskView'}}
        next_cursor:
          oneOf:
            - {type: string, minLength: 1, maxLength: 4096}
            - {type: 'null'}
    TaskProfileOption:
      type: object
      additionalProperties: false
      required: [ref, name, revision, digest, capabilities, real_model_allowed]
      properties:
        ref: {type: string, minLength: 1, maxLength: 256}
        name: {type: string, minLength: 1, maxLength: 256}
        revision: {$ref: '#/components/schemas/RevisionString'}
        digest: {type: string, pattern: '^[a-f0-9]{64}$'}
        capabilities: {type: array, maxItems: 128, uniqueItems: true, items: {type: string, minLength: 1, maxLength: 128}}
        real_model_allowed: {type: boolean}
        max_explore_concurrency:
          description: Published RuntimeProfile ceiling and default for Task Explore concurrency. Null or absent means this option does not expose that runtime capability; model profiles do not use it.
          oneOf:
            - {type: integer, minimum: 1, maximum: 256}
            - {type: 'null'}
    ReadinessCheck:
      type: object
      additionalProperties: false
      required: [id, layer, status, reason_code, observed_at, remediation_owner, message]
      properties:
        id: {type: string, minLength: 1, maxLength: 128}
        layer: {type: string, enum: [identity, profile, budget, model, runtime, target, material]}
        status: {type: string, enum: [pass, fail, unknown, not_applicable]}
        reason_code: {type: string, pattern: '^[a-z][a-z0-9_]{0,127}$'}
        observed_at: {type: string, format: date-time}
        evidence_ref:
          oneOf:
            - {type: string, minLength: 1, maxLength: 256}
            - {type: 'null'}
        remediation_owner: {type: string, enum: [user, application, gateway, infrastructure]}
        message: {type: string, minLength: 1, maxLength: 1024}
    TaskOptions:
      type: object
      additionalProperties: false
      required: [project_id, model_profiles, runtime_profiles, missing]
      properties:
        project_id: {type: string, minLength: 1, maxLength: 256}
        model_profiles: {type: array, maxItems: 256, items: {$ref: '#/components/schemas/TaskProfileOption'}}
        runtime_profiles: {type: array, maxItems: 256, items: {$ref: '#/components/schemas/TaskProfileOption'}}
        missing: {type: array, maxItems: 64, items: {$ref: '#/components/schemas/ReadinessCheck'}}
    ReadinessReport:
      type: object
      additionalProperties: false
      required: [task_id, definition_digest, observed_at, can_request_start, checks]
      properties:
        task_id: {type: string, minLength: 1, maxLength: 256}
        definition_digest: {type: string, pattern: '^[a-f0-9]{64}$'}
        observed_at: {type: string, format: date-time}
        can_request_start: {type: boolean}
        checks: {type: array, maxItems: 64, items: {$ref: '#/components/schemas/ReadinessCheck'}}
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message, request_id, retryable, details]
      properties:
        code: {$ref: '#/components/schemas/ErrorCode'}
        message: {type: string, minLength: 1, maxLength: 8192}
        request_id: {type: string, minLength: 1, maxLength: 256}
        retryable: {type: boolean}
        details:
          type: object
          additionalProperties: true
...